refactor: `TabBar`의 렌더링 계층을 변경
