+++ frontend/src/components/Template/CustomerTemplate/BottomTabBar/index.tsx
@@ -17,25 +17,17 @@ const BOTTOM_TABS = [
 const BottomTabBar = () => {
   const location = useLocation();
 
-  const isTabBarVisible = [
-    ROUTER_PATH.couponList,
-    ROUTER_PATH.rewardList,
-    ROUTER_PATH.myPage,
-  ].includes(location.pathname as RouterPath);
-
   return (
-    isTabBarVisible && (
-      <TabBarContainer>
-        {BOTTOM_TABS.map((tab, index) => (
-          <TapBarItem key={index} $isSelected={tab.path.includes(location.pathname as RouterPath)}>
-            <Link to={tab.path[0]}>
-              {tab.icon}
-              {tab.label}
-            </Link>
-          </TapBarItem>
-        ))}
-      </TabBarContainer>
-    )
+    <TabBarContainer>
+      {BOTTOM_TABS.map((tab, index) => (
+        <TapBarItem key={index} $isSelected={tab.path.includes(location.pathname as RouterPath)}>
+          <Link to={tab.path[0]}>
+            {tab.icon}
+            {tab.label}
+          </Link>
+        </TapBarItem>
+      ))}
+    </TabBarContainer>
   );
 };
