+++ frontend/src/components/Template/CustomerTemplate/index.tsx
@@ -3,15 +3,22 @@ import { BaseCustomerTemplate, ContentContainer } from './style';
 import BottomTabBar from './BottomTabBar';
 import { useLocation } from 'react-router-dom';
 import { ROUTER_PATH } from '../../../constants';
+import { RouterPath } from '../../../types';
 
 const CustomerTemplate = ({ children }: PropsWithChildren) => {
   const location = useLocation();
+
+  const isTabBarVisible = [
+    ROUTER_PATH.couponList,
+    ROUTER_PATH.rewardList,
+    ROUTER_PATH.myPage,
+  ].includes(location.pathname as RouterPath);
 
   return (
     <BaseCustomerTemplate $isHome={location.pathname === ROUTER_PATH.couponList}>
       <ContentContainer>
         {children}
-        <BottomTabBar />
+        {isTabBarVisible && <BottomTabBar />}
       </ContentContainer>
     </BaseCustomerTemplate>
   );
